refactor to add to emitters
